+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Prettier automatic tailwind classes sorting.
+- Pass the entire message in LS input and let the LLM chose articles.
 
 ### Removed
+- `order` column in `messages` table + fixed.
 - SQLite support.
 
 ## Added
@@ -18,7 +20,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Possibility to interrupt streaming.
 - Filter and sort get_thread and get_thread/messages.
 - Plots for the SCS simulations.
+- New semantic route.
 
+### Fixed
+- Literature search points to right url.
+
+## [v0.4.7] - 15.04.2025
+
+### Fixed
+- Empty groups from keycloak.
 
 ## [v0.4.6]